# [E2E] Fix grep spec filtering